run_model: Take --color to disable conversion to greyscale
The Test.ai model was trained on greyscale images (many of them black
and white), but the Appium plugin doesn't currently convert the image to
greyscale. This allows us to copy that behaviour to measure what effect
it has on accuracy.

Note that this doesn't make any difference on Test.ai's training images,
because they are already stored on disk as greyscale images. Well, it
makes a speed difference: It cuts off 70% of the execution time. I
suppose the greyscale conversion is the slowest part of the script. It's
got to be faster in numpy or something.